Add progressive strategy

+++ src/com/hashcoders/noexception/Intersection.java
@@ -4,13 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Intersection {
-	float latitude,longitude;
+	double latitude,longitude;
 	int i;
 	
 	List<Road> incoming = new ArrayList<Road>();
 	List<Road> outgoing = new ArrayList<Road>();
 
-	public Intersection(float latitude, float longitude, int i) {
+	public Intersection(double latitude, double longitude, int i) {
 		this.latitude = latitude;
 		this.longitude = longitude;
 		this.i = i;
